server/requirements.txt
banana-dev
sanic
requests
python-dotenv